refactor(service.UrlService): Remove static UrlUtils class to use a injectable service
